Avoid strncpy for CMessageHeader::pchCommand
The warning is spurious, as we exactly want the truncation behaviour here (no zero padding in pchCommand when the input has length exactly 12), but it seems hard to avoid. The reason is that strncopy is designed to operate on C strings, and our target isn't exactly a C string.

This was changed in bitcoin PR #18367 to a simple loop that assigns to every byte in the output once, but if desirable, a solution with strlen + memcpy + memset is also possible.